# backend/app/web/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "股票FinCore AI API",
    "description": "FinCore AI Stock Analysis System REST API",
    "version": "2.1.0"
  },
  "host": "localhost:8888",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "APIKeyHeader": {
      "type": "apiKey",
      "name": "X-API-KEY",
      "in": "header"
    },
    "HMACSignature": {
      "type": "apiKey",
      "name": "X-HMAC-SIGNATURE",
      "in": "header"
    }
  },
  "paths": {
    "/analyze": {
      "post": {
        "summary": "Analyze a single stock",
        "description": "Performs a comprehensive analysis of a single stock, including technical indicators, fundamental data, and AI-driven insights. This is a synchronous operation.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "APIKeyHeader": [],
            "HMACSignature": []
          }
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AnalysisRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful analysis",
            "schema": {
              "$ref": "#/definitions/AnalysisResult"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/api/start_stock_analysis": {
      "post": {
        "summary": "Start a stock analysis task",
        "description": "Starts an asynchronous stock analysis task. The client should poll the status endpoint to get the result.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "APIKeyHeader": [],
            "HMACSignature": []
          }
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AnalysisRequest"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Task accepted",
            "schema": {
              "$ref": "#/definitions/TaskCreationResponse"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/api/stock_data": {
      "get": {
        "summary": "Get historical stock data",
        "description": "Retrieves historical price data and technical indicators for a given stock.",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "APIKeyHeader": [],
            "HMACSignature": []
          }
        ],
        "parameters": [
          {
            "name": "stock_code",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "The stock code, e.g., '600519'"
          },
          {
            "name": "market_type",
            "in": "query",
            "required": false,
            "type": "string",
            "default": "A",
            "description": "Market type: 'A', 'HK', or 'US'"
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "1m",
              "3m",
              "6m",
              "1y"
            ],
            "default": "1y",
            "description": "Time period for the data"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of stock data"
          },
          "400": {
            "description": "Invalid parameters",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "AnalysisRequest": {
      "type": "object",
      "properties": {
        "stock_code": {
          "type": "string",
          "example": "600519"
        },
        "market_type": {
          "type": "string",
          "example": "A"
        }
      },
      "required": ["stock_code"]
    },
    "TaskCreationResponse": {
        "type": "object",
        "properties": {
            "task_id": {
                "type": "string"
            }
        }
    },
...
    "Error": {
        "type": "object",
        "properties": {
            "error": {
                "type": "string"
            }
        }
    },
...
    "AnalysisResult": {
      "type": "object",
      "properties": {
        "score": {
          "type": "number"
        },
        "recommendation": {
          "type": "string"
        }
      }
    }
  }
}